collapse initialization into a single tx
